server/requirements.txt
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
google-genai
spacy
ffmpeg-python
# Optional: uncomment if you need server-side Whisper transcription
# openai-whisper
